fixed error and slight refactor
